requirements.txt
torch
torchvision
matplotlib
streamlit
opencv-python
pillow
